requirements.txt
comet-ml==3.12.0
numpy~=1.19.5
tensorflow~=2.3.0
tensorflow-gpu~=2.3.0
transformers~=3.0.2
scikit-learn==0.24.2
tensorflow-text~=2.3.0
tf-models-official~=2.3.0
pandas~=1.3.0